Fix bug with Settings.dll. Bump version to 2.6

+++ src/VisualStudio/Settings/SettingsManagerBase.cs
@@ -48,7 +48,7 @@ namespace NuGet.VisualStudio
         protected bool ReadBool(string settingsRoot, string property, bool defaultValue = false)
         {
             var userSettingsStore = _settingsManager.Value.GetReadOnlySettingsStore();
-            if (userSettingsStore.CollectionExists(settingsRoot))
+            if (userSettingsStore != null && userSettingsStore.CollectionExists(settingsRoot))
             {
                 return userSettingsStore.GetBoolean(settingsRoot, property, defaultValue);
             }
@@ -61,13 +61,16 @@ namespace NuGet.VisualStudio
         protected void WriteBool(string settingsRoot, string property, bool value)
         {
             IWritableSettingsStore userSettingsStore = GetWritableSettingsStore(settingsRoot);
-            userSettingsStore.SetBoolean(settingsRoot, property, value);
+            if (userSettingsStore != null)
+            {
+                userSettingsStore.SetBoolean(settingsRoot, property, value);
+            }
         }
 
         protected int ReadInt32(string settingsRoot, string property, int defaultValue = 0)
         {
             var userSettingsStore = _settingsManager.Value.GetReadOnlySettingsStore();
-            if (userSettingsStore.CollectionExists(settingsRoot))
+            if (userSettingsStore != null && userSettingsStore.CollectionExists(settingsRoot))
             {
                 return userSettingsStore.GetInt32(settingsRoot, property, defaultValue);
             }
@@ -80,13 +83,16 @@ namespace NuGet.VisualStudio
         protected void WriteInt32(string settingsRoot, string property, int value)
         {
             IWritableSettingsStore userSettingsStore = GetWritableSettingsStore(settingsRoot);
-            userSettingsStore.SetInt32(settingsRoot, property, value);
+            if (userSettingsStore != null)
+            {
+                userSettingsStore.SetInt32(settingsRoot, property, value);
+            }
         }
 
         protected string ReadString(string settingsRoot, string property, string defaultValue = "")
         {
             var userSettingsStore = _settingsManager.Value.GetReadOnlySettingsStore();
-            if (userSettingsStore.CollectionExists(settingsRoot))
+            if (userSettingsStore != null && userSettingsStore.CollectionExists(settingsRoot))
             {
                 return userSettingsStore.GetString(settingsRoot, property, defaultValue);
             }
@@ -99,7 +105,7 @@ namespace NuGet.VisualStudio
         protected string[] ReadStrings(string settingsRoot, string[] properties, string defaultValue = "")
         {
             var userSettingsStore = _settingsManager.Value.GetReadOnlySettingsStore();
-            if (userSettingsStore.CollectionExists(settingsRoot))
+            if (userSettingsStore != null && userSettingsStore.CollectionExists(settingsRoot))
             {
                 string[] values = new string[properties.Length];
                 for (int i = 0; i < values.Length; i++)
@@ -117,7 +123,14 @@ namespace NuGet.VisualStudio
         protected bool DeleteProperty(string settingsRoot, string property)
         {
             IWritableSettingsStore userSettingsStore = GetWritableSettingsStore(settingsRoot);
-            return userSettingsStore.DeleteProperty(settingsRoot, property);
+            if (userSettingsStore != null)
+            {
+                return userSettingsStore.DeleteProperty(settingsRoot, property);
+            }
+            else
+            {
+                return false;
+            }
         }
 
         protected void WriteStrings(string settingsRoot, string[] properties, string[] values)
@@ -125,23 +138,28 @@ namespace NuGet.VisualStudio
             Debug.Assert(properties.Length == values.Length);
 
             IWritableSettingsStore userSettingsStore = GetWritableSettingsStore(settingsRoot);
-            for (int i = 0; i < properties.Length; i++)
+            if (userSettingsStore != null)
             {
-                userSettingsStore.SetString(settingsRoot, properties[i], values[i]);
+                for (int i = 0; i < properties.Length; i++)
+                {
+                    userSettingsStore.SetString(settingsRoot, properties[i], values[i]);
+                }
             }
         }
 
         protected void WriteString(string settingsRoot, string property, string value)
         {
             IWritableSettingsStore userSettingsStore = GetWritableSettingsStore(settingsRoot);
-
-            userSettingsStore.SetString(settingsRoot, property, value);
+            if (userSettingsStore != null)
+            {
+                userSettingsStore.SetString(settingsRoot, property, value);
+            }
         }
 
         protected void ClearAllSettings(string settingsRoot)
         {
             IWritableSettingsStore userSettingsStore = _settingsManager.Value.GetWritableSettingsStore();
-            if (userSettingsStore.CollectionExists(settingsRoot))
+            if (userSettingsStore != null && userSettingsStore.CollectionExists(settingsRoot))
             {
                 userSettingsStore.DeleteCollection(settingsRoot);
             }
@@ -150,7 +168,7 @@ namespace NuGet.VisualStudio
         private IWritableSettingsStore GetWritableSettingsStore(string settingsRoot)
         {
             IWritableSettingsStore userSettingsStore = _settingsManager.Value.GetWritableSettingsStore();
-            if (!userSettingsStore.CollectionExists(settingsRoot))
+            if (userSettingsStore != null && !userSettingsStore.CollectionExists(settingsRoot))
             {
                 userSettingsStore.CreateCollection(settingsRoot);
             }
